+++ noodles-csi/src/async/reader.rs
@@ -179,6 +179,13 @@ async fn read_bins<R>(
 where
     R: AsyncRead + Unpin,
 {
+    fn duplicate_bin_error(id: usize) -> io::Result<(IndexMap<usize, Bin>, Option<Metadata>)> {
+        Err(io::Error::new(
+            io::ErrorKind::InvalidData,
+            format!("duplicate bin ID: {id}"),
+        ))
+    }
+
     let n_bin = reader.read_i32_le().await.and_then(|n| {
         usize::try_from(n).map_err(|e| io::Error::new(io::ErrorKind::InvalidData, e))
     })?;
@@ -199,12 +206,18 @@ where
             .map(bgzf::VirtualPosition::from)?;
 
         if id == metadata_id {
-            metadata = read_metadata(reader).await.map(Some)?;
+            let m = read_metadata(reader).await?;
+
+            if metadata.replace(m).is_some() {
+                return duplicate_bin_error(id);
+            }
         } else {
             let chunks = read_chunks(reader).await?;
             let bin = Bin::new(loffset, chunks);
-            // TODO: Check for duplicates.
-            bins.insert(id, bin);
+
+            if bins.insert(id, bin).is_some() {
+                return duplicate_bin_error(id);
+            }
         }
     }